# Critical Infrastructure Enhancement: Implementation of MC Hammer-Inspired ActiveRecord Method Alias