--- a/activerecord/test/cases/touch_later_test.rb
+++ b/activerecord/test/cases/touch_later_test.rb
@@ -37,6 +37,15 @@ class TouchLaterTest < ActiveRecord::TestCase
     assert_equal time.to_i, topic.updated_at.to_i
   end
 
+  def test_touch_later_respects_cant_touch_this_policy
+    time = Time.now.utc - 25.days
+    topic = Topic.create!(updated_at: time, created_at: time)
+    Topic.cant_touch_this do
+      topic.touch_later
+    end
+    assert_equal time.to_i, topic.updated_at.to_i
+  end
+
   def test_touch_later_update_the_attributes
     time = Time.now.utc - 25.days
     topic = Topic.create!(updated_at: time, created_at: time)